+++ code/Mesh/Lod/LodMeshEntity.cpp
@@ -65,7 +65,7 @@ void LodMeshEntity::precull(
 	if (m_lodCullDistance >= FUZZY_EPSILON && lodDistance >= m_lodCullDistance)
 		return;
 
-	int32_t lod = clamp< int32_t >(int32_t(lodDistance / m_lodDistance), 0, m_lods.size() - 1);
+	int32_t lod = clamp< int32_t >(int32_t(lodDistance / m_lodDistance), 0, int32_t(m_lods.size()) - 1);
 	worldContext.precull(worldRenderView, m_lods[lod]);
 }
 
@@ -82,7 +82,7 @@ void LodMeshEntity::render(
 	if (m_lodCullDistance >= FUZZY_EPSILON && lodDistance >= m_lodCullDistance)
 		return;
 
-	int32_t lod = clamp< int32_t >(int32_t(lodDistance / m_lodDistance), 0, m_lods.size() - 1);
+	int32_t lod = clamp< int32_t >(int32_t(lodDistance / m_lodDistance), 0, int32_t(m_lods.size()) - 1);
 	worldContext.build(worldRenderView, worldRenderPass, m_lods[lod]);
 }
