+++ code/Ui/Custom/ToolBar/ToolBarDropDown.cpp
@@ -154,7 +154,7 @@ void ToolBarDropDown::buttonDown(ToolBar* toolBar, MouseEvent* mouseEvent)
 	PopupMenu menu;
 	if (menu.create())
 	{
-		for (size_t i = 0; i < m_items.size(); ++i)
+		for (uint32_t i = 0; i < uint32_t(m_items.size()); ++i)
 			menu.add(new MenuItem(Command(i), m_items[i]));
 		
 		Ref< MenuItem > selectedItem = menu.show(toolBar, m_menuPosition);
